pyproject.toml
[tool.poetry]
name = "drf-api-permissions"
version = "0.1.0"
description = ""
authors = ["Brendan Welzien <brendanwelzien@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.1.4"
djangorestframework = "^3.12.2"
psycopg2-binary = "^2.8.6"

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
